app/otel.py
import logging

from config import settings

logger = logging.getLogger("sag_rag.otel")

def setup_tracing(app=None):
    if not settings.otel_enabled:
        return
    try:
        from opentelemetry import trace
        from opentelemetry.sdk.resources import Resource
        from opentelemetry.sdk.trace import TracerProvider
        from opentelemetry.sdk.trace.export import BatchSpanProcessor
        from opentelemetry.exporter.otlp.proto.grpc.trace_exporter import OTLPSpanExporter
    except Exception as exc:
        logger.warning("OpenTelemetry not available: %s", exc)
        return
    resource = Resource.create({"service.name": settings.otel_service_name})
    provider = TracerProvider(resource=resource)
    exporter = OTLPSpanExporter(endpoint=settings.otel_exporter_otlp_endpoint, insecure=True)
    provider.add_span_processor(BatchSpanProcessor(exporter))
    trace.set_tracer_provider(provider)
    try:
        from opentelemetry.instrumentation.fastapi import FastAPIInstrumentor
        from opentelemetry.instrumentation.requests import RequestsInstrumentor
        if app is not None:
            FastAPIInstrumentor().instrument_app(app)
        RequestsInstrumentor().instrument()
    except Exception:
        pass
    logger.info("OpenTelemetry tracing enabled")
