app/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic>=2
pydantic-settings

httpx
aiohttp
qdrant-client
neo4j
elasticsearch==8.12.0

sentence-transformers==2.6.0

transformers==4.39.3

torch==2.2.1
accelerate==0.28.0

spacy==3.7.2

tqdm

# < NEW FIXED VERSION >
google-generativeai==0.7.2

pytest==8.1.1

opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-instrumentation-requests==0.46b0
